requirements.txt
Flask
authlib
python-dotenv
gunicorn